% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdczx.plot.R
\name{scdczx.plot}
\alias{scdczx.plot}
\title{Graphing with data}
\usage{
scdczx.plot(infect)
}
\arguments{
\item{infect}{List of the reasult.}
}
\value{
plots.
}
\description{
Using Data for Mapping
}
\examples{
scdczx.plot(data)

}
