% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdczx.R
\name{scdczx}
\alias{scdczx}
\title{Data preprocessing for cdc}
\usage{
scdczx(data)
}
\arguments{
\item{szk}{Infectious disease reporting card data}
}
\value{
List of the reasult
}
\description{
Preprocessing of infectious disease reporting card data
}
\examples{
scdczx(data)

}
